fix: 31 bug fixes from comprehensive codebase review
Critical fixes:
- i16 overflow in profit calculation (matched * 100) changed to i64
- TOCTOU race in circuit breaker cooldown check (read->write lock)
- Silent circuit breaker trips now logged with reason
- Premature WS connected flags removed (set before handshake)
- Atomic file writes in state_writer (temp+rename pattern)
- Cache miss no longer silently guesses lowercase mapping

Execution engine:
- profit_cents/expected_profit_cents widened to i64 across types
- Failed execution errors logged instead of silently swallowed
- Unwind failures upgraded from warn to error level
- Trade log writes converted from sync to async

Risk & position management:
- record_success now routes P&L to correct position side by ArbType
- Position monitor wired to RiskManager.record_exit (fixes ghost positions)
- RiskManager.update_mode returns early when Stopped (sticky floor)
- RiskConfig.from_env validates positive percentages
- Circuit breaker background tasks (cooldown check, daily P&L reset)
- Signal trader: failed exits no longer record phantom P&L
- Signal trader: maker orders with 0 fills no longer create positions

Market maker:
- on_fill uses action field instead of position existence for buy/sell
- QuoteRequest carries side field for correct reprice direction
- sell_all uses entry_price/2 instead of 1 cent
- Stuck position double-sell failures alert via Telegram

WebSocket & API:
- Kalshi delta: zero-quantity best-bid removal detected and cleared
- Polymarket: upward ask moves no longer filtered out
- Polymarket CLOB: neg_risk_cache uses tokio::sync::RwLock (async-safe)
- Polymarket CLOB: HTTP status checked before JSON parse
- Polymarket CLOB: orderID extraction fails loudly instead of "unknown"
- Kalshi channel-full on exec request now warns instead of silent drop
- Signing debug logs removed (security)

Other:
- Discovery errors propagated instead of silently dropped
- Daily P&L in state writer reads from circuit breaker instead of hardcoded 0
- SizeCents doc comment corrected (max ~$655, not ~$655k)
- Integration tests updated for new ArbType parameter and MarketPair fields

All 139 tests pass. 17 files changed, +332 -116.

+++ src/telegram.rs
@@ -166,7 +166,7 @@ impl TelegramClient {
         kalshi_filled: i64,
         poly_filled: i64,
         total_cost_cents: i64,
-        profit_cents: i16,
+        profit_cents: i64,
         poly_slug: &str,
     ) {
         let msg = format!(
